+++ src/lib/components/layout/Frame.js
@@ -28,6 +28,8 @@ var Frame = React.createClass({
 			className: React.PropTypes.string,
 			style: React.PropTypes.object,
 			bgColor: React.PropTypes.string,
+			width: numberOrStringType,
+			height: numberOrStringType,
 			align: React.PropTypes.oneOf(["left", "center", "right"]),
 			valign: React.PropTypes.oneOf(["top", "middle", "bottom"])
 		})